Rename ../generated/schema to ../generated/AnchorSchema
